// src/types/hr-management/employee-management.d.ts
interface IEmployeeUnit {
    children?: IEmployeeUnit[];
    unitId: string;
    unitName: string;
    isSmallest: boolean;
    parentId: number;
    totalMainWorkingMember: number;
    unitTypeId: number;
    unitTypeName: string;
    unitTypeLevel: string;
}

export interface IEmployeeIndexes {
    companies: { companyId: string; companyName: string }[];
    units: IEmployeeUnit[];
    positions: { positionId: string; positionName: string }[];
    grades: string[];
    resignationStatuses: { statusId: string; statusName: string }[];
    hrPositionId: { positionId: string; positionName: string }[];
    genders: { genderId: string; genderName: string }[];
    maritalStatuses: { maritalStatusId: string; maritalStatusName: string }[];
    nationalities: { nationalityId: string; nationalityName: string; isActive: false }[];
    statuses: { statusId: string; statusName: string }[];
    educationNames: string[];
    locationNames: string[];
    mainProjects: { projectId: string; projectName: string }[];
}
// ...
export interface IAddNewEmployeeForm {
    firstName: string;
    lastName: string;
    birthday: string;
    genderId: number;
    martialStatusId: number;
    nationalityId: number;
    contactAddress: string;
    mobilePhone: string;
    badgeId: string;
    workEmail: string;
    positionId: number;
    grade: number;
    joinDate: string;
    projectId: number;
    experienceBefore: number;
    personalEmail: string;
    homePhone: string;
    isDegree: boolean;
}